comit

+++ src/main/java/com/zr/test/demo/service/impl/AppServiceImpl.java
@@ -9,6 +9,8 @@ import com.zr.test.demo.component.exception.CustomException;
 import com.zr.test.demo.config.enums.ErrorCode;
 import com.zr.test.demo.model.dto.AppDTO;
 import com.zr.test.demo.model.dto.AppQueryDTO;
+import com.zr.test.demo.model.dto.StatusDTO;
+import com.zr.test.demo.model.entity.AfterCourse;
 import com.zr.test.demo.model.entity.App;
 import com.zr.test.demo.dao.AppMapper;
 import com.zr.test.demo.model.entity.AppCategory;
@@ -24,6 +26,7 @@ import com.zr.test.demo.util.TimeUtil;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.*;
@@ -87,7 +90,7 @@ public class AppServiceImpl extends ServiceImpl<AppMapper, App> implements IAppS
         }
         List<AppVO> res = new ArrayList<>();
         List<AppCategory> categories = appCategoryService.getBaseMapper().selectList(null);
-        Map<String, String> map = categories.stream().collect(Collectors.toMap(e -> "" + e.getId(), AppCategory::getName));
+        Map<String, String> map = categories.stream().collect(Collectors.toMap(e -> e.getId().toString(), AppCategory::getName));
         page.getRecords().forEach(app -> {
             AppVO vo = new AppVO();
             BeanUtils.copyProperties(app, vo);
@@ -95,13 +98,13 @@ public class AppServiceImpl extends ServiceImpl<AppMapper, App> implements IAppS
             if (file != null) {
                 vo.setLogo(Optional.ofNullable(file.getFilePath()).orElse(""));
             }
-            vo.setType(Optional.ofNullable(map.get("" + app.getType())).orElse(""));
-            vo.setSubject(Optional.ofNullable(map.get("" + app.getSubject())).orElse(""));
-            vo.setPlatform(Optional.ofNullable(map.get("" + app.getPlatform())).orElse(""));
+            vo.setType(Optional.ofNullable(map.get(app.getType().toString())).orElse(""));
+            vo.setSubject(Optional.ofNullable(map.get(app.getSubject().toString())).orElse(""));
+            vo.setPlatform(Optional.ofNullable(map.get(app.getPlatform().toString())).orElse(""));
             String[] arr = app.getTags().split(",");
             StringBuilder sb = new StringBuilder();
             for (String idStr : arr) {
-                String temp = map.get("" + idStr);
+                String temp = map.get(idStr);
                 if (StringUtil.isEmpty(temp)) {
                     continue;
                 }
@@ -121,29 +124,74 @@ public class AppServiceImpl extends ServiceImpl<AppMapper, App> implements IAppS
 
     @Override
     public Result<AppOneVO> queryOne(Long id, HttpServletRequest request) {
-        App app=this.getBaseMapper().selectById(id);
-        AppOneVO vo=new AppOneVO();
-        BeanUtils.copyProperties(app,vo);
+        App app = this.getBaseMapper().selectById(id);
+        AppOneVO vo = new AppOneVO();
+        BeanUtils.copyProperties(app, vo);
         vo.setImg(FileUtil.getBase64FilePath(fileRouterService.selectPath(app.getImg())));
+        vo.setImgId(app.getImg());
         vo.setLogo(FileUtil.getBase64FilePath(fileRouterService.selectPath(app.getLogo())));
+        vo.setLogoId(app.getLogo());
         vo.setTags(ListUtil.stringToList(app.getTags()));
         return Result.success(vo);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public Result<Object> updateByDto(AppDTO dto, HttpServletRequest request) {
-        if(dto.getId()==null){
+        if (dto.getId() == null) {
             throw new CustomException(ErrorCode.SYS_PARAM_INNER_ERR);
         }
-        App app=new App();
-        BeanUtils.copyProperties(dto,app);
+        App oldApp = this.getBaseMapper().selectById(dto.getId() );
+        if(!oldApp.getLogo().equals(dto.getId())){
+            fileRouterService.deleteOldFile(oldApp.getLogo());
+        }
+        if(!oldApp.getImg().equals(dto.getImg())){
+            fileRouterService.deleteOldFile(oldApp.getImg());
+        }
+        App app = new App();
+        BeanUtils.copyProperties(dto, app);
         app.setTags(ListUtil.listToString(dto.getTags()));
         return Result.success(this.getBaseMapper().updateById(app));
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public Result<Object> delete(Long id, HttpServletRequest request) {
+        App oldApp = this.getBaseMapper().selectById(id );
+        if (oldApp == null) {
+            return Result.success(0);
+        }
+        fileRouterService.deleteOldFile(oldApp.getLogo());
+        fileRouterService.deleteOldFile(oldApp.getImg());
         return Result.success(this.getBaseMapper().deleteById(id));
     }
+
+    @Override
+    public Result<Object> updateStatus(StatusDTO dto, HttpServletRequest request) {
+        App course = new App();
+        course.setId(dto.getId());
+        course.setStatus(dto.getStatus());
+        return Result.success(this.getBaseMapper().updateById(course));
+    }
+
+    @Override
+    public Result<List<AppOneVO>> queryAppName(String name, HttpServletRequest request) {
+        QueryWrapper<App> queryWrapper = new QueryWrapper<>();
+        queryWrapper.like("name", name);
+        List<App> list=this.getBaseMapper().selectList(queryWrapper);
+        if(ListUtil.isEmpty(list)){
+            return Result.success(new ArrayList<>());
+        }
+        List<AppOneVO> voList=new ArrayList<>();
+        list.forEach(l->{
+            AppOneVO vo=new AppOneVO();
+            vo.setId(l.getId());
+            vo.setName(l.getName());
+            vo.setLogo(FileUtil.getBase64FilePath(fileRouterService.selectPath(l.getLogo())));
+            voList.add(vo);
+        });
+        return Result.success(voList);
+    }
+
 
 }
